+++ src/main/kotlin/com/quarkbyte/recoveryapp_api/service/casecsj/SinistroService.kt
@@ -45,9 +45,8 @@ class SinistroService(private val repository: SinistroRepository) {
 
     @Throws(ResourceNotFoundException::class)
     fun delete(id: UUID): ResponseEntity<*> {
-        val saved = repository.findById(id)
-            .orElseThrow { ResourceNotFoundException("None cases founded") }!!
+        if (repository.findById(id).isPresent)
         repository.deleteById(id)
-        return ResponseEntity.ok("$saved deleted sucessfully")
+        return ResponseEntity.ok("deleted successfully")
     }
 }
