+++ src/main/kotlin/com/quarkbyte/recoveryapp_api/service/customer/BondsmanService.kt
@@ -55,9 +55,8 @@ class BondsmanService(
 
     @Throws(ResourceNotFoundException::class)
     fun delete(id: UUID): ResponseEntity<*> {
-        val saved = repository.findById(id)
-            .orElseThrow { ResourceNotFoundException("None address founded") }!!
+        if (repository.findById(id).isPresent)
         repository.deleteById(id)
-        return ResponseEntity.ok("$saved deleted successfully")
+        return ResponseEntity.ok("deleted successfully")
     }
 }
